Fix the bug of buffer update.
Add two arguments (`method_by_name`, `method_by_type`) in `set_fix_method`. The fixed-point methods of different parameters/buffers could be specified by module name or module types, to support another option of accumulating BN running_mean/var: 1) accumulating and forwarding with floating-point arithmetic during training, 2) only quantizing the running mean/var during the test forwarding process. This option might be needed when the bitwidth of running mean and variance is small.

Bug record: When using `BatchNorm2d_fix` along with DataParallel, the network would not work in eval mode. The running means and vars are not correctly accumulated, *because*, after the quantized buffer tensor is in-place updated by batchnorm operation, it is referenced by the replication of the module object in the DataParallel scenario, this reference is of no use. Change: copy the updated buffer's data back to the module's buffer tensor (the buffer tensor on the primary gpu would be updated).